+++ mira/metamodel/__init__.py
@@ -3,5 +3,6 @@ from .io import *
 from .templates import *
 from .template_model import *
 from .comparison import *
+from .schema import *
 from .search import *
 from .ops import *
